requirements.txt
fastapi==0.73.0
uvicorn==0.17.5
pillow==9.0.1
pytesseract==0.3.8
requests==2.27.1
beautifulsoup4==4.10.0
tabula-py==2.3.0